:bee: fix: update dist files #32 それに伴いもろもろ変更。
