+++ src/main.js
@@ -1,8 +1,1 @@
 
-// import '../lib/main';
-import { RaitingStar } from '../lib/main';
-
-new RaitingStar({
-  // filledColor: 'pink',
-  // emptyColor: 'red',
-});
