+++ test/index.html
@@ -3,9 +3,9 @@
 
 <head>
   <meta charset="UTF-8" />
-  <link rel="icon" type="image/svg+xml" href="/vite.svg" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Test Page</title>
+  <link rel="stylesheet" href="./../dist/ratingStar.css">
 </head>
 
 <body>
@@ -22,10 +22,22 @@
   <div class="rs">3/5</div>
   <div class="rs">0.2/5</div>
   <div class="rs">0.5/1</div>
+  <div class="test">5/10</div>
   <!-- <div class="rs">0/0</div> -->
   <!-- <div class="rs">2/1</div> -->
 
-  <script type="module" src="/src/main.js"></script>
+  <!-- <script src="./../dist/ratingStar.umd.js"></script>
+  <script>
+    const ratingStar = new RatingStar();
+    ratingStar.init();
+  </script> -->
+  <!-- <script type="module" src="./../lib/main.ts"></script>
+  <script>
+    const ratingStar = new RatingStar();
+    ratingStar.init();
+  </script> -->
+  <script type="module" src="./main.js"></script>
+
 </body>
 
 </html>
