+++ test/main.js
@@ -1,0 +1,4 @@
+import { RatingStar } from '../lib/ratingStar.ts';
+
+const ratingStar = new RatingStar();
+ratingStar.init({});
