--- test/test.js
@@ -1,8 +1,0 @@
-import './test.css';
-// import '../lib/main';
-import { RaitingStar } from '../lib/main';
-
-new RaitingStar({
-  filledColor: 'pink',
-  emptyColor: 'red',
-});
