+++ vite.config.ts
@@ -4,12 +4,13 @@ export default defineConfig({
   build: {
     lib: {
       entry: 'lib/main.ts',
-      name: 'ratingStar',
-      fileName: 'ratingStar',
+      name: 'RatingStar',
+      formats: ['umd'],
+      fileName: (format) => `ratingStar.${format}.js`,
     },
     rollupOptions: {
       output: {
-        assetFileNames: 'raitingStar.css',
+        assetFileNames: 'ratingStar.css',
       },
     },
   },
